# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Domicilia Core API
  version: 0.1.0
  description: |
    Contrato de la API de domicilia-core. Es la FUENTE DE VERDAD para cualquier cliente (el frontend, el agente de IA):
    los clientes tipados se generan de este archivo.

    Convenciones:
    - Identificadores: UUID como texto.
    - Errores: `application/problem+json` (RFC 9457) con `detail` y `request_id`.
    - Listados que crecen (conversaciones, mensajes, contactos): paginación por CURSOR (`limit` + `cursor`); la respuesta
      trae `next_cursor` (null en la última página). El cursor es opaco.
    - Fechas: RFC 3339 en UTC.
    - Cambios parciales (PATCH): un campo ausente no se toca; `null` deja el campo en blanco.

    ESTADO: este contrato cubre por ahora el módulo de WhatsApp (bandejas, contactos, conversaciones y mensajes) y su
    webhook, el módulo de catálogo (categorías, productos, variantes y modificadores), el de pedidos (carrito y su
    máquina de estados), el de pagos (cobro por ePayco y su webhook) y el de promociones (cupones de descuento). El
    resto de la API (usuarios, organizaciones, planes, roles) se documenta aparte; una prueba
    verifica que lo documentado aquí es exactamente lo que el servidor sirve, y valida las respuestas reales contra
    estos esquemas.
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
  /v1/organizations/{org_id}/categories:
    parameters:
      - $ref: '#/components/parameters/OrgID'
    get:
      tags: [Catálogo]
      operationId: listCategories
      summary: Lista las categorías del menú
      description: Requiere `org.catalog.read`. Incluye activas e inactivas — quien administra el menú necesita ver ambas para poder reactivar una.
      responses:
        '200':
          description: Categorías
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Category' } }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [Catálogo]
      operationId: createCategory
      summary: Crea una categoría
      description: Requiere `org.catalog.manage`. El nombre es único por organización entre las activas.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCategoryRequest' }
      responses:
        '201':
          description: Categoría creada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/organizations/{org_id}/categories/{category_id}:
    parameters:
      - $ref: '#/components/parameters/OrgID'
      - $ref: '#/components/parameters/CategoryID'
    patch:
      tags: [Catálogo]
      operationId: updateCategory
      summary: Cambia una categoría (parcial)
      description: Requiere `org.catalog.manage`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateCategoryRequest' }
      responses:
        '200':
          description: Categoría actualizada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/public/products:
    get:
      tags: [Catálogo]
      operationId: publicProductFeed
      summary: Feed de Inicio — productos de varias organizaciones a la vez
      description: |
        Sin sesión. Solo organizaciones activas y productos activos Y publicados a este canal (`channels` incluye
        `"ecommerce"` — un producto no aparece aquí solo por existir, la organización tiene que publicarlo) con al
        menos una variante activa (sin eso no hay nada que agregar al carrito). `category` filtra por el NOMBRE de
        categoría (sin distinguir mayúsculas) — las categorías son por organización, no una taxonomía compartida,
        así que es una coincidencia de texto, no un id. Para configurar un producto antes de agregarlo al carrito,
        pedir su detalle completo (`GET /public/organizations/{org_id}/products/{product_id}`).
      parameters:
        - name: organization_id
          in: query
          description: Acota al menú de UN restaurante por id — lo usa su propia vitrina si ya lo conoce.
          schema: { type: string, format: uuid }
        - name: organization_slug
          in: query
          description: Lo mismo que organization_id, pero por slug — la vitrina de un restaurante solo conoce el slug de su URL.
          schema: { type: string }
        - name: q
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: Página del feed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FeedProductPage' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/public/organizations/{org_id}/products/{product_id}:
    parameters:
      - $ref: '#/components/parameters/OrgID'
      - $ref: '#/components/parameters/ProductID'
    get:
      tags: [Catálogo]
      operationId: publicGetProductDetail
      summary: Detalle de un producto, para configurarlo antes de agregarlo al carrito
      description: Sin sesión. La organización debe existir y estar activa; el producto, activo.
      responses:
        '200':
          description: Producto con variantes y modificadores completos
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductDetail' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/organizations/{org_id}/products:
    parameters:
      - $ref: '#/components/parameters/OrgID'
    get:
      tags: [Catálogo]
      operationId: listProducts
      summary: Lista productos (paginado)
      description: Requiere `org.catalog.read`.
      parameters:
        - name: category_id
          in: query
          schema: { type: string, format: uuid }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 30 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: Página de productos
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductPage' }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [Catálogo]
      operationId: createProduct
      summary: Crea un producto
      description: |
        Requiere `org.catalog.manage`. Exige al menos una variante (el precio vive ahí, nunca en el producto). Si
        ninguna variante viene marcada `is_default`, se marca la primera automáticamente.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateProductRequest' }
      responses:
        '201':
          description: Producto creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/organizations/{org_id}/products/{product_id}:
    parameters:
      - $ref: '#/components/parameters/OrgID'
      - $ref: '#/components/parameters/ProductID'
    get:
      tags: [Catálogo]
      operationId: getProduct
      summary: Lee un producto
      description: Requiere `org.catalog.read`.
      responses:
        '200':
          description: Producto
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        default: { $ref: '#/components/responses/Problem' }
    patch:
      tags: [Catálogo]
      operationId: updateProduct
      summary: Cambia un producto (parcial)
      description: |
        Requiere `org.catalog.manage`. `variants` y `modifier_group_ids` ausentes no se tocan; presentes (incluso
        vacíos) REEMPLAZAN la lista completa: una variante existente que no venga en la lista nueva se desactiva —
        nunca se borra, porque un pedido futuro podría referenciarla.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateProductRequest' }
      responses:
        '200':
          description: Producto actualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/organizations/{org_id}/products/{product_id}/image:
    parameters:
      - $ref: '#/components/parameters/OrgID'
      - $ref: '#/components/parameters/ProductID'
    post:
      tags: [Catálogo]
      operationId: uploadProductImage
      summary: Sube la foto de un producto
      description: |
        Requiere `org.catalog.manage`. El tipo real se detecta por el CONTENIDO del archivo (nunca por su nombre
        ni por el Content-Type que mande el cliente): solo JPEG, PNG, WEBP o GIF, hasta 5 MB. Reemplaza cualquier
        foto anterior. Responde 503 si este servidor no tiene el almacenamiento de fotos configurado
        (CORE_MEDIA_STORAGE_ACCOUNT).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Producto con la foto actualizada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/organizations/{org_id}/modifier-groups:
    parameters:
      - $ref: '#/components/parameters/OrgID'
    get:
      tags: [Catálogo]
      operationId: listModifierGroups
      summary: Lista los grupos de modificadores de la organización
      description: Requiere `org.catalog.read`. No pagina — un menú real tiene pocas decenas de grupos como mucho.
      responses:
        '200':
          description: Grupos de modificadores
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/ModifierGroup' } }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [Catálogo]
      operationId: createModifierGroup
      summary: Crea un grupo de modificadores
      description: Requiere `org.catalog.manage`. Se reutiliza entre productos (p. ej. "Extras de pizza" se define una vez).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateModifierGroupRequest' }
      responses:
        '201':
          description: Grupo creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ModifierGroup' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/organizations/{org_id}/modifier-groups/{modifier_group_id}:
    parameters:
      - $ref: '#/components/parameters/OrgID'
      - $ref: '#/components/parameters/ModifierGroupID'
    get:
      tags: [Catálogo]
      operationId: getModifierGroup
      summary: Lee un grupo de modificadores
      description: Requiere `org.catalog.read`.
      responses:
        '200':
          description: Grupo de modificadores
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ModifierGroup' }
        default: { $ref: '#/components/responses/Problem' }
    patch:
      tags: [Catálogo]
      operationId: updateModifierGroup
      summary: Cambia un grupo de modificadores (parcial)
      description: 'Requiere `org.catalog.manage`. `options` sigue el mismo reemplazo-completo que `variants` de un producto.'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateModifierGroupRequest' }
      responses:
        '200':
          description: Grupo actualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ModifierGroup' }
        default: { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT de auth-domicilia (GoTrue).
  parameters:
    OrgID:
      name: org_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    CategoryID:
      name: category_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    ProductID:
      name: product_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    ModifierGroupID:
      name: modifier_group_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Problem:
      description: |
        Error. Códigos habituales: 401 sin credenciales; 402 el plan no incluye la función; 403 sin permiso o organización
        suspendida; 404 no existe (o archivada); 409 choca con el estado actual; 422 entrada inválida; 502 falló Meta;
        503 función sin configurar.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, example: 'about:blank' }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        request_id: { type: string }
    Category:
      type: object
      required: [id, organization_id, name, position, is_active, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        name: { type: string }
        position: { type: integer }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateCategoryRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        position: { type: integer, default: 0 }
    UpdateCategoryRequest:
      type: object
      description: Un campo ausente no se toca. Ningún campo admite null.
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        position: { type: integer }
        is_active: { type: boolean }
...
    Product:
      type: object
      required: [id, organization_id, category_id, name, description, image_url, position, is_active, variants, modifier_group_ids, ingredients, channels, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        category_id: { type: string, format: uuid, nullable: true }
        name: { type: string }
        description: { type: string, nullable: true }
        image_url: { type: string, nullable: true }
        position: { type: integer }
        is_active: { type: boolean }
        variants: { type: array, items: { $ref: '#/components/schemas/Variant' } }
        modifier_group_ids: { type: array, items: { type: string, format: uuid } }
        ingredients: { type: array, items: { type: string }, description: 'Texto libre de la organización, sin validar contra nada.' }
        channels:
          type: array
          items: { type: string, enum: [ecommerce, whatsapp] }
          description: 'Dónde se publica. Vacío por omisión: un producto nuevo NO aparece en el feed público solo por estar activo — hay que publicarlo a propósito.'
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateProductRequest:
      type: object
      description: 'Exige al menos una variante: el precio vive ahí, nunca en el producto. Sin channels, el producto se crea sin publicar en ningún lado.'
      required: [name, variants]
      properties:
        category_id: { type: string, format: uuid, nullable: true }
        name: { type: string, minLength: 1, maxLength: 150 }
        description: { type: string, maxLength: 1000, nullable: true }
        image_url: { type: string, nullable: true }
        position: { type: integer, default: 0 }
        variants: { type: array, minItems: 1, maxItems: 30, items: { $ref: '#/components/schemas/VariantInput' } }
        modifier_group_ids: { type: array, maxItems: 20, items: { type: string, format: uuid } }
        ingredients: { type: array, maxItems: 50, items: { type: string, maxLength: 60 } }
        channels: { type: array, items: { type: string, enum: [ecommerce, whatsapp] } }
    UpdateProductRequest:
      type: object
      description: |
        Un campo ausente no se toca. En `category_id`, `description` e `image_url`, `null` explícito lo deja en
        blanco. `variants`, `modifier_group_ids`, `ingredients` y `channels` ausentes no se tocan; presentes
        (incluso `[]`) REEMPLAZAN la lista completa — una variante que no venga se desactiva, nunca se borra;
        `channels: []` despublica el producto de todos los canales sin desactivarlo.
      properties:
        category_id: { type: string, format: uuid, nullable: true }
        name: { type: string, minLength: 1, maxLength: 150 }
        description: { type: string, maxLength: 1000, nullable: true }
        image_url: { type: string, nullable: true }
        position: { type: integer }
        is_active: { type: boolean }
        variants: { type: array, maxItems: 30, items: { $ref: '#/components/schemas/VariantInput' } }
        modifier_group_ids: { type: array, maxItems: 20, items: { type: string, format: uuid } }
        ingredients: { type: array, maxItems: 50, items: { type: string, maxLength: 60 } }
        channels: { type: array, items: { type: string, enum: [ecommerce, whatsapp] } }
    ProductPage:
      type: object
      required: [items, total, limit, offset]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Product' } }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
    FeedProduct:
      type: object
      description: |
        Un producto tal como aparece en el feed público de Inicio (varias organizaciones a la vez). Liviano a
        propósito: para configurarlo y agregarlo al carrito, pedir el detalle completo
        (GET /public/organizations/{org_id}/products/{product_id}).
      required: [id, organization_id, organization_name, organization_slug, organization_logo_url, category_id, category_name, name, description, image_url, min_price_cents]
      properties:
        id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        organization_name: { type: string }
        organization_slug: { type: string }
        organization_logo_url: { type: string, nullable: true }
        category_id: { type: string, format: uuid, nullable: true }
        category_name: { type: string, nullable: true }
        name: { type: string }
        description: { type: string, nullable: true }
        image_url: { type: string, nullable: true }
        min_price_cents: { type: integer, description: Precio de la variante activa más barata — "desde $X" en la tarjeta. }
    FeedProductPage:
      type: object
      required: [items, total, limit, offset]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/FeedProduct' } }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
    ProductDetail:
      type: object
      description: Un producto con sus variantes y grupos de modificadores completos — lo que hace falta para configurarlo antes de agregarlo al carrito.
      allOf:
        - $ref: '#/components/schemas/Product'
        - type: object
          required: [modifier_groups]
          properties:
            modifier_groups: { type: array, items: { $ref: '#/components/schemas/ModifierGroup' } }
...
    OptionInput:
      type: object
      required: [name, price_delta_cents]
      properties:
        id: { type: string, format: uuid, nullable: true }
        name: { type: string, minLength: 1, maxLength: 100 }
        price_delta_cents: { type: integer, minimum: 0, maximum: 100000000 }
    ModifierGroup:
      type: object
      description: 'Se reutiliza entre productos. min_select >= 1 es lo que hace al grupo obligatorio — no hay un campo "required" aparte.'
      required: [id, organization_id, name, selection_type, min_select, max_select, is_active, options, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        name: { type: string }
        selection_type: { type: string, enum: [single, multiple] }
        min_select: { type: integer, minimum: 0 }
        max_select: { type: integer, minimum: 1 }
        is_active: { type: boolean }
        options: { type: array, items: { $ref: '#/components/schemas/ModifierOption' } }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateModifierGroupRequest:
      type: object
      required: [name, selection_type, options]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        selection_type: { type: string, enum: [single, multiple] }
        min_select: { type: integer, minimum: 0, default: 0 }
        max_select: { type: integer, minimum: 1, default: 1 }
        options: { type: array, minItems: 1, maxItems: 50, items: { $ref: '#/components/schemas/OptionInput' } }
    UpdateModifierGroupRequest:
      type: object
      description: |
        Un campo ausente no se toca. `selection_type`/`min_select`/`max_select` se validan juntos contra los valores
        vigentes (cambiar solo uno podría dejar la combinación inconsistente). `options` ausente no se toca; presente
        (incluso `[]`) REEMPLAZA la lista completa.
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        selection_type: { type: string, enum: [single, multiple] }
        min_select: { type: integer, minimum: 0 }
        max_select: { type: integer, minimum: 1 }
        is_active: { type: boolean }
        options: { type: array, maxItems: 50, items: { $ref: '#/components/schemas/OptionInput' } }
...